Add Mascota and Microchip service implementations
Se implementaron las clases MascotaServiceImpl y MicrochipServiceImpl para gestionar la lógica de negocio y las operaciones CRUD para las entidades Mascota y Microchip. Se actualizó la interfaz GenericService para usar el tipo Long en lugar de int para los identificadores de entidad, manteniendo así la coherencia. MascotaServiceImpl gestiona las operaciones transaccionales que involucran tanto a Mascota como a Microchip, garantizando la integridad de los datos para las relaciones uno a uno.

+++ Main/src/service/GenericService.java
@@ -1,8 +1,4 @@
 /*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
-/*
  * Interfaz genérica de servicios de negocio.
  * Define las operaciones CRUD básicas que debe ofrecer cualquier Service.
  */
@@ -13,12 +9,16 @@ import java.util.List;
 public interface GenericService<T> {
 // insterta una nueva entidad en el sistema
     void crear(T entidad) throws Exception;
+    
 //Identificador de la entidad por su ID
-    T leer(int id) throws Exception;
+    T leer(Long id) throws Exception;
+    
 //Retorna todas las entidades del sistema
     List<T> leerTodos() throws Exception;
+    
 //Actualiza una entidad existente
     void actualizar(T entidad) throws Exception;
+    
 //Elimina una entidad por su ID
-    void eliminar(int id) throws Exception;
+    void eliminar(Long id) throws Exception;
 }
